Changed implementation of change detection.

+++ madmex/lcc/transform/elliptic.py
@@ -86,4 +86,4 @@ class Transform(TransformBase):
         change_classification[change_classification==1]=0
         change_classification[change_classification==-1]=1
 
-        return change_classification.astype(np.int8)
+        return change_classification
